requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.2
django-ckeditor==6.0.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-paypal==1.0.0
filelock==3.0.12
future==0.18.2
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
#matplotlib==3.2.1
mccabe==0.6.1
MouseInfo==0.1.3
#numpy==1.18.2
#pandas==1.0.3
pefile==2019.4.18
Pillow==7.0.0
pipenv==2020.11.15
playsound==1.2.2
#psycopg2==2.8.4
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyInstaller==3.6
pylint==2.4.4
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-firebase==1.2
PyTweening==1.0.3
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.9
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.11.2
